Temporarily show the menubar when pressing alt if hidden.
I don't know how to connect to menuBar()::focusOutEvent to hide the
menubar when the menu closes. Since you need to override that function,
I'd have to extend QMenuBar (or probably KMenuBar) and whatever
function initializes the menubar too.

So for now, it will remain visible until you tap Alt again with the
menu already closed.